--- app/src/main/java/project/seo/pictureviewer/ui/main/MainActivity.kt
@@ -1,16 +1,0 @@
-package project.seo.pictureviewer.ui.main
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import dagger.hilt.android.AndroidEntryPoint
-import project.seo.pictureviewer.databinding.ActivityMainBinding
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-    }
-}
