--- app/src/main/java/project/seo/pictureviewer/ui/main/MainListViewHolder.kt
@@ -1,25 +1,0 @@
-package project.seo.pictureviewer.ui.main
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import project.seo.pictureviewer.data.Picture
-import project.seo.pictureviewer.databinding.ListItemBinding
-
-class MainListViewHolder(
-    private val binding: ListItemBinding,
-) : RecyclerView.ViewHolder(binding.root) {
-    fun bind(pictureData: Picture) {
-        binding.data = pictureData
-    }
-
-    companion object {
-        operator fun invoke(parent: ViewGroup): MainListViewHolder {
-            return MainListViewHolder(
-                ListItemBinding.inflate(
-                    LayoutInflater.from(parent.context), parent, false
-                )
-            )
-        }
-    }
-}
