--- app/src/main/java/project/seo/pictureviewer/utils/ViewKtx.kt
@@ -1,9 +1,0 @@
-package project.seo.pictureviewer.utils
-
-import android.view.View
-import androidx.lifecycle.LifecycleCoroutineScope
-import androidx.lifecycle.coroutineScope
-import androidx.lifecycle.findViewTreeLifecycleOwner
-
-val View.coroutineScope: LifecycleCoroutineScope?
-    get() = findViewTreeLifecycleOwner()?.lifecycle?.coroutineScope
