+++ presentation/src/main/java/project/seo/pictureviewer/presentation/ui/main/MainActivity.kt
@@ -1,0 +1,16 @@
+package project.seo.pictureviewer.presentation.ui.main
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import dagger.hilt.android.AndroidEntryPoint
+import project.seo.pictureviewer.presentation.databinding.ActivityMainBinding
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+    }
+}
