+++ presentation/src/main/java/project/seo/pictureviewer/presentation/ui/main/MainFragment.kt
@@ -1,0 +1,46 @@
+package project.seo.pictureviewer.presentation.ui.main
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
+import androidx.navigation.NavController
+import androidx.navigation.Navigation
+import dagger.hilt.android.AndroidEntryPoint
+import project.seo.pictureviewer.presentation.databinding.FragmentMainBinding
+import project.seo.pictureviewer.presentation.utils.RecyclerViewItemDecorator
+
+@AndroidEntryPoint
+class MainFragment : Fragment() {
+    private lateinit var binding: FragmentMainBinding
+    private val viewModel: MainViewModel by viewModels()
+    private lateinit var navigator: NavController
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View {
+        super.onCreateView(inflater, container, savedInstanceState)
+        binding = FragmentMainBinding.inflate(layoutInflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        navigator = Navigation.findNavController(view)
+
+        with(binding) {
+            vm = viewModel
+            lifecycleOwner = viewLifecycleOwner
+            recyclerView.adapter = MainListAdapter { pictureId ->
+                navigator.navigate(
+                    MainFragmentDirections.actionMainFragmentToDetailFragment(pictureId)
+                )
+            }
+            recyclerView.addItemDecoration(RecyclerViewItemDecorator())
+        }
+    }
+}
