+++ presentation/src/main/java/project/seo/pictureviewer/presentation/utils/ViewKtx.kt
@@ -1,0 +1,9 @@
+package project.seo.pictureviewer.presentation.utils
+
+import android.view.View
+import androidx.lifecycle.LifecycleCoroutineScope
+import androidx.lifecycle.coroutineScope
+import androidx.lifecycle.findViewTreeLifecycleOwner
+
+val View.coroutineScope: LifecycleCoroutineScope?
+    get() = findViewTreeLifecycleOwner()?.lifecycle?.coroutineScope
