+++ presentation/src/main/res/drawable/back_arrow.png
