+++ presentation/src/main/res/drawable/default_image.png
