+++ presentation/src/main/res/drawable/next_arrow.png
