+++ presentation/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/mainFragment">
+
+    <fragment
+        android:id="@+id/mainFragment"
+        android:name="project.seo.pictureviewer.presentation.ui.main.MainFragment"
+        android:label="MainFragment">
+        <action
+            android:id="@+id/action_mainFragment_to_detailFragment"
+            app:destination="@id/detailFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/detailFragment"
+        android:name="project.seo.pictureviewer.presentation.ui.detail.DetailFragment"
+        android:label="DetailFragment">
+        <argument
+            android:name="pictureId"
+            app:argType="integer" />
+    </fragment>
+</navigation>
